chage limit type long to int

+++ hugegraph-client/src/main/java/com/baidu/hugegraph/api/traverser/CrosspointsAPI.java
@@ -43,7 +43,7 @@ public class CrosspointsAPI extends TraversersAPI {
     public List<Path> get(Object sourceId, Object targetId,
                           Direction direction, String label,
                           int maxDepth, long degree,
-                          long capacity, long limit) {
+                          long capacity, int limit) {
         String source = GraphAPI.formatVertexId(sourceId, false);
         String target = GraphAPI.formatVertexId(targetId, false);
